+++ source/extensions/lightspeed.trex.hotkeys/lightspeed/trex/hotkeys/app.py
@@ -73,6 +73,8 @@ class TrexHotkeyEvent(HotkeyEvent):
     F = KeyCombination(carb.input.KeyboardInput.F)
     # to unselect all selected objects
     ESC = KeyCombination(carb.input.KeyboardInput.ESCAPE)
+    # teleport selection as in a viewport
+    CTRL_T = KeyCombination(carb.input.KeyboardInput.T, modifiers=carb.input.KEYBOARD_MODIFIER_FLAG_CONTROL)
 
 
 class HotkeyManager: